+++ sources/base/Enums.swift
@@ -22,9 +22,7 @@ enum Settings {
     static let unreleasedSets = "unreleasedSets"
     static let displayWelcome = "displayWelcome"
     static let collectionNumberBadge = "collectionNumberBadge"
-
-    
-
+    static let compactList = "compactList"
 
 }
 enum SheetType  {